Renamed toc component file to table of contents
